Moved weapon bag into its own file
