+++ Assets/Weaponary/WeaponBag.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 7143f2bdbda683e4e947d178be4f5612
+timeCreated: 1484090583
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
